+++ src/JosephGuadagno.Broadcasting.Web/Controllers/LinkedInController.cs
@@ -1,4 +1,5 @@
 using System.Text.Json;
+using JosephGuadagno.Broadcasting.Domain.Constants;
 using JosephGuadagno.Broadcasting.Data.KeyVault.Interfaces;
 using JosephGuadagno.Broadcasting.Web.Models;
 using JosephGuadagno.Broadcasting.Web.Models.LinkedIn;
@@ -11,7 +12,7 @@ namespace JosephGuadagno.Broadcasting.Web.Controllers;
 /// <summary>
 /// Handles the LinkedIn OAuth2 flow
 /// </summary>
-[Authorize(Policy = "RequireContributor")]
+[Authorize(Policy = AuthorizationPolicyNames.RequireContributor)]
 public class LinkedInController : Controller
 {
     private readonly HttpClient _httpClient;