+++ src/JosephGuadagno.Broadcasting.Web/Controllers/SchedulesController.cs
@@ -13,7 +13,7 @@ namespace JosephGuadagno.Broadcasting.Web.Controllers;
 /// <summary>
 /// The controller for the schedules
 /// </summary>
-[Authorize(Policy = "RequireContributor")]
+[Authorize(Policy = AuthorizationPolicyNames.RequireContributor)]
 public class SchedulesController : Controller
 {
     private readonly IScheduledItemService _scheduledItemService;